refactor(flashcard): extract CardStyles and AutoSizedText into own files
FlashcardView.swift reduced from 1001 to 818 lines by promoting:
- CardTransitionDirection, BouncyButtonStyle, HeartButtonStyle,
  CardActionButton → CardStyles.swift
- AutoSizedText (formerly nested) → AutoSizedText.swift
